+++ packages/gestalt/src/index.js
@@ -41,7 +41,6 @@ import Text from './Text.js';
 import TextArea from './TextArea.js';
 import TextField from './TextField.js';
 import Toast from './Toast.js';
-import Tooltip from './Tooltip.js';
 import Touchable from './Touchable.js';
 import Video from './Video.js';
 
@@ -88,7 +87,6 @@ export {
   TextArea,
   TextField,
   Toast,
-  Tooltip,
   Touchable,
   Video,
 };